Rename file hero-banner.vue to HeroBanner.vue
